Update ipa-xsampa.csv
I've been using this file for one of my projects and I noticed a few discrepancies, which I ironed out. Doesn't look like anything depends on the descriptions so I thought I would share in case it's useful for others.

On another note, it seems like epitran relies this file. I thought that instead of having it in two places, epitran could read it from the panphon module. If that's useful, let me know and I can open a pull request there.